Boost: libs: pass BOOST_USE_WINAPI_VERSION as define (#583)
When `add_compile_definitions()` is used in a cmake toolchain file the
directory property COMPILE_DEFINITIONS is set. Use this property to pass
those definitions to the `boost-user.jam` file.

This is introduced to pass the `BOOST_USE_WINAPI_VERSION` as define
to the bjam build command.

This fixes Boost.Log usage when the project is explicitly built for Windows 8
or newer. Boost.Log builds per default build for Windows 7, which breaks
the build because the boost-header will be used in a Windows 8 context,
but the Windows 8 specific functions are not compiled into the library.

The changelog for Boost 1.78 mentions `BOOST_USE_WINAPI_VERSION` for
`Atomic` and `Log` libraries https://www.boost.org/users/history/version_1_78_0.html

> ### Atomic:
> - On Windows, corrected discrepancy between `atomic-type::always_has_native_wait_notify` and the corresponding capability macros when targeting Windows 8 or later. The library will now directly use `WaitOnAddress` and related APIs from public headers and therefore require user to link with `synchronization.lib` if the user requires Windows 8 or later by defining `BOOST_USE_WINAPI_VERSION`, `_WIN32_WINNT` or similar macros. The library is linked automatically on compilers that support auto-linking (e.g. MSVC).

> ### Log:
> - On Windows, when building the library for Windows 8 or later, the library will use `nt62` tag in the version namespace to denote the target OS ABI. For example, the version namespace could be named as `v2_mt_nt62`. This name will be part of all symbols exported by the library. Use the `BOOST_USE_WINAPI_VERSION` macro consistenly when building Boost and your code to request the minimum target Windows version.

As an explicit override the CMake variable `BOOST_USE_WINAPI_VERSION`
can be set (either by CMAKE_ARGS or in a toolchain file) to bjam using the
`define=` bjam commandline. This overrides the define set by `add_compile_definitions()`.

The `define=` usage is documented at
https://www.boost.org/doc/libs/1_79_0/libs/log/doc/html/log/installation/config.html

The `BOOST_USE_WINAPI_VERSION` define is documented for WinAPI and Log:
- https://www.boost.org/doc/libs/1_79_0/libs/winapi/doc/html/winapi/config.html
- https://www.boost.org/doc/libs/1_79_0/libs/log/doc/html/log/installation/config.html

Also push the internal schema version number to trigger a rebuild, as the
schema is changed.