docs: Add SHA1 calculation onliner and fix sums (#724)
This commit improves the documentation of the prcess for adding new
packages. It replaces two shell commands for SHA1 calculation with
one using a pipe. This allows you not to delete the downloaded
archive from the file system later.

The commit also corrects the SHA1 sums shown in the example.